% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cybersickness_row.R
\docType{data}
\name{cybersickness_row}
\alias{cybersickness_row}
\title{Raw Physiological Measurements for Cybersickness Study}
\format{
An object of class \code{data.frame} with 25893 rows and 17 columns.
}
\usage{
cybersickness_row
}
\description{
Raw data from a cybersickness study including heart rate, heart rate variability, and their percent changes.
}
\details{
A data frame with 25893 rows and 17 columns:
Epoch: Time interval index.
HR: Heart Rate.
PC_HR: Percent change in Heart Rate.
HR_MIN: Minimum heart rate in the epoch.
HR_MAX: Maximum heart rate in the epoch.
HRV: Heart Rate Variability.
PC_HRV: Percent change in HRV.
HRV_MIN: Minimum HRV in the epoch.
HRV_MAX: Maximum HRV in the epoch.
...: Other columns include similar physiological signals.
}
\keyword{datasets}
