% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.savvySh.R
\name{coef.savvySh_model}
\alias{coef.savvySh_model}
\title{Extract Coefficients From a Slab and Shrinkage Model Object}
\usage{
\method{coef}{savvySh_model}(object, estimator = NULL, ...)
}
\arguments{
\item{object}{A fitted model object of class \code{savvySh_model}.}

\item{estimator}{The specific estimator to use for extracting coefficients. For example, \code{"St"}, \code{"DSh"}, etc.
This must match one of the estimators in the fitted \code{savvySh_model} object.
Defaults to the first available estimator if not specified.}

\item{...}{Additional arguments passed to the \code{predict} method.}
}
\value{
A named vector of coefficients for the specified estimator.
}
\description{
Extract the estimated coefficients from a fitted Slab and Shrinkage linear regression model object.
}
\seealso{
\code{\link{predict.savvySh_model}}, \code{\link{savvySh}}
}
\author{
Ziwei Chen, Vali Asimit, Marina Anca Cidota, Jennifer Asimit\cr
Maintainer: Ziwei Chen <ziwei.chen.3@bayes.city.ac.uk>
}
