% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cybersickness_row.R
\docType{data}
\name{cybersickness_row}
\alias{cybersickness_row}
\title{Raw Physiological Measurements for Cybersickness Study}
\format{
A data frame with 25893 rows and 17 columns:
\describe{
  \item{Epoch}{Time interval index.}
  \item{HR}{Heart Rate.}
  \item{PC_HR}{Percent change in Heart Rate.}
  \item{HR_MIN}{Minimum heart rate in the epoch.}
  \item{HR_MAX}{Maximum heart rate in the epoch.}
  \item{HRV}{Heart Rate Variability.}
  \item{PC_HRV}{Percent change in HRV.}
  \item{HRV_MIN}{Minimum HRV in the epoch.}
  \item{HRV_MAX}{Maximum HRV in the epoch.}
  \item{...}{Other columns include similar physiological signals.}
}
}
\usage{
cybersickness_row
}
\description{
Raw data from a cybersickness study including heart rate, heart rate variability, and their percent changes.
}
\keyword{datasets}
