% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.savvSh.R
\name{predict.savvySh_model}
\alias{predict.savvySh_model}
\title{Predict for Slab and Shrinkage Linear Regression Estimation}
\usage{
\method{predict}{savvySh_model}(
  object,
  newx = NULL,
  type = c("response", "coefficients"),
  estimator = NULL,
  ...
)
}
\arguments{
\item{object}{Fitted \code{"savvySh_model"} object returned by \code{savvySh}.}

\item{newx}{Matrix of new data for which predictions are to be made. Must have the same number of columns as the training data.
This argument is required for \code{type = "response"}.}

\item{type}{Type of prediction required. Can be \code{"response"} (fitted values) or \code{"coefficients"}.
Defaults to \code{"response"}.}

\item{estimator}{The specific estimator to use for prediction. For example, \code{"St"}, \code{"DSh"} if \code{model_class = "Multiplicative"}, etc.
This must match one of the estimators in the fitted \code{savvySh_model} object.
Defaults to the first available estimator if not specified.}

\item{...}{Additional arguments.}
}
\value{
Predicted values for \code{type = "response"} or coefficients for \code{type = "coefficients"}.
}
\description{
This function predicts fitted values or coefficients for a fitted Slab and Shrinkage linear regression model
created using the \code{savvySh} function.
}
\seealso{
\code{\link{savvySh}}
}
\author{
Ziwei Chen, Vali Asimit, Marina Anca Cidota, Jennifer Asimit\cr
Maintainer: Ziwei Chen <ziwei.chen.3@citystgeorges.ac.uk>
}
