% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.savvySh.R
\name{print.savvySh_model}
\alias{print.savvySh_model}
\title{Print a Slab and Shrinkage Linear Regression Model Object}
\usage{
\method{print}{savvySh_model}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{Fitted \code{"savvySh_model"} object.}

\item{digits}{Significant digits in the printout.}

\item{...}{Additional print arguments.}
}
\value{
Invisibly returns a data frame summarizing the model (number of coefficients and lambda values, if any).
}
\description{
Print a summary of the Slab and Shrinkage linear regression model object.
}
\details{
The function prints the call that produced the \code{savvySh_model} object and a summary of the model,
including the selected model class, the number of non-zero coefficients for each estimator, and
if applicable, the optimal lambda value.
}
\seealso{
\code{\link{savvySh}}
}
\author{
Ziwei Chen, Vali Asimit, Marina Anca Cidota, Jennifer Asimit\cr
Maintainer: Ziwei Chen <ziwei.chen.3@bayes.city.ac.uk>
}
