% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.savvySh.R
\name{summary.savvySh_model}
\alias{summary.savvySh_model}
\title{Summary of a Slab and Shrinkage Linear Regression Estimation with Statistics}
\usage{
\method{summary}{savvySh_model}(object, ...)
}
\arguments{
\item{object}{A fitted model object of class \code{savvySh_model}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
This function provides a summary of the fitted \code{savvySh_model}, displaying
the estimated coefficients, residuals, standard errors, R-squared, and F-statistics.
}
\seealso{
\code{\link{savvySh}}
}
\author{
Ziwei Chen, Vali Asimit, Marina Anca Cidota, Jennifer Asimit\cr
Maintainer: Ziwei Chen <ziwei.chen.3@citystgeorges.ac.uk>
}
